+++ FABLECRAFT_MODERN_STACK/eslint.config.js
@@ -48,6 +48,14 @@ export default [
         window: 'readonly',
         navigator: 'readonly',
         console: 'readonly',
+        localStorage: 'readonly',
+        sessionStorage: 'readonly',
+        crypto: 'readonly',
+        setTimeout: 'readonly',
+        clearTimeout: 'readonly',
+        setInterval: 'readonly',
+        clearInterval: 'readonly',
+        React: 'readonly',
         // Node globals for build files
         process: 'readonly',
         Buffer: 'readonly',
