+++ FABLECRAFT_MODERN_STACK/src/features-modern/auth/components/protected-route.tsx
@@ -119,35 +119,4 @@ export const RequireAuth: React.FC<RequireAuthProps> = ({
   return <>{children}</>
 }
 
-/**
- * AUTHENTICATION GUARD HOOK
- * 
- * Custom hook for imperative authentication checks in components
- * 
- * Usage:
- * ```tsx
- * const { isAuthenticated, isLoading, redirectToLogin } = useAuthGuard()
- * 
- * if (isLoading) return <Loading />
- * if (!isAuthenticated) return redirectToLogin()
- * ```
- */
-export const useAuthGuard = (redirectTo: string = '/login') => {
-  const isAuthenticated = useIsAuthenticated()
-  const isLoading = useAuthLoading()
-  const isInitialized = useAuthInitialized()
-  
-  const redirectToLogin = () => (
-    <Navigate to={redirectTo} replace />
-  )
-  
-  return {
-    isAuthenticated,
-    isLoading,
-    isInitialized,
-    redirectToLogin,
-    needsRedirect: !isLoading && isInitialized && !isAuthenticated,
-  }
-}
-
 export default ProtectedRoute
