+++ FABLECRAFT_MODERN_STACK/src/features-modern/dashboard/components/dashboard-layout.tsx
@@ -1,4 +1,4 @@
-import { ReactNode } from 'react'
+import React, { ReactNode } from 'react'
 
 /**
  * DASHBOARD LAYOUT PROPS INTERFACE
