+++ FABLECRAFT_MODERN_STACK/tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/App.tsx","./src/main.tsx","./src/app/providers/theme-context.ts","./src/app/providers/theme-provider.tsx","./src/app/providers/use-theme.ts","./src/components/ui/badge-variants.ts","./src/components/ui/badge.tsx","./src/components/ui/button-variants.ts","./src/components/ui/button.test.tsx","./src/components/ui/button.tsx","./src/components/ui/card.tsx","./src/components/ui/dropdown-menu.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/features-modern/auth/components/login-form.tsx","./src/features-modern/auth/components/protected-route.tsx","./src/features-modern/auth/pages/login-page.tsx","./src/features-modern/auth/stores/auth-store.ts","./src/features-modern/dashboard/components/adaptive-header.tsx","./src/features-modern/dashboard/components/dashboard-layout.tsx","./src/features-modern/dashboard/stores/dashboard-store.ts","./src/features-modern/landing/index.ts","./src/features-modern/landing/landing-page.tsx","./src/features-modern/landing/components/cta-section.tsx","./src/features-modern/landing/components/feature-cards.tsx","./src/features-modern/landing/components/footer-section.tsx","./src/features-modern/landing/components/hero-section.tsx","./src/features-modern/landing/components/pricing-section.tsx","./src/features-modern/landing/components/process-steps.tsx","./src/features-modern/landing/components/testimonials-section.tsx","./src/features-modern/theme/components/theme-toggle.tsx","./src/features-modern/theme/components/__tests__/theme-toggle.test.tsx","./src/features-modern/theme/config/theme-config.ts","./src/lib/utils.ts","./src/migration/tracker.ts","./src/pages/HomePage.tsx","./src/pages/NotFoundPage.tsx","./src/shared/components/badge-with-dot.tsx","./src/shared/components/error-boundary.tsx","./src/test/integration.test.ts","./src/test/setup.ts","./src/test/mocks/handlers.ts","./src/test/mocks/server.ts","./vite.config.ts"],"version":"5.8.3"}
+{"root":["./src/App.tsx","./src/main.tsx","./src/app/providers/theme-context.ts","./src/app/providers/theme-provider.tsx","./src/app/providers/use-theme.ts","./src/components/ui/badge-variants.ts","./src/components/ui/badge.tsx","./src/components/ui/button-variants.ts","./src/components/ui/button.test.tsx","./src/components/ui/button.tsx","./src/components/ui/card.tsx","./src/components/ui/dropdown-menu.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/features-modern/auth/components/login-form.tsx","./src/features-modern/auth/components/protected-route.tsx","./src/features-modern/auth/hooks/use-auth-guard.tsx","./src/features-modern/auth/pages/login-page.tsx","./src/features-modern/auth/stores/auth-store.ts","./src/features-modern/dashboard/components/adaptive-header.tsx","./src/features-modern/dashboard/components/dashboard-layout.tsx","./src/features-modern/dashboard/stores/dashboard-store.ts","./src/features-modern/landing/index.ts","./src/features-modern/landing/landing-page.tsx","./src/features-modern/landing/components/cta-section.tsx","./src/features-modern/landing/components/feature-cards.tsx","./src/features-modern/landing/components/footer-section.tsx","./src/features-modern/landing/components/hero-section.tsx","./src/features-modern/landing/components/pricing-section.tsx","./src/features-modern/landing/components/process-steps.tsx","./src/features-modern/landing/components/testimonials-section.tsx","./src/features-modern/theme/components/theme-toggle.tsx","./src/features-modern/theme/components/__tests__/theme-toggle.test.tsx","./src/features-modern/theme/config/theme-config.ts","./src/lib/utils.ts","./src/migration/tracker.ts","./src/pages/HomePage.tsx","./src/pages/NotFoundPage.tsx","./src/shared/components/badge-with-dot.tsx","./src/shared/components/error-boundary.tsx","./src/test/integration.test.ts","./src/test/setup.ts","./src/test/mocks/handlers.ts","./src/test/mocks/server.ts","./vite.config.ts"],"version":"5.8.3"}
