+++ src/features-modern/auth/components/protected-route.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import { Navigate, Outlet } from 'react-router-dom'
 import { useAuthStore, useIsAuthenticated, useAuthLoading, useAuthInitialized } from '../stores/auth-store'
