+++ src/features-modern/auth/pages/login-page.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import { useNavigate } from 'react-router-dom'
 import { useEffect } from 'react'
 import { useIsAuthenticated } from '../stores/auth-store'
